requirements.txt
pyzbar
Pillow
pyinstaller
